+++ config/failed-jobs-notifier.php
@@ -7,14 +7,14 @@ return [
     /*
      * The number of days limit from which jobs will be considered as stuck.
      */
-    'daysLimit' => 2,
+    'daysLimit'    => 2,
 
     /*
      * The notifiable to which the notification will be sent. The default
      * notifiable will use the mail and slack configuration specified
      * in this config file.
      */
-    'notifiable' => \Okipa\LaravelFailedJobsNotifier\Notifiable::class,
+    'notifiable'   => \Okipa\LaravelFailedJobsNotifier\Notifiable::class,
 
     /*
      * The notification that will be sent when stuck jobs are detected.
@@ -25,9 +25,9 @@ return [
     * You can pass a boolean or a callable to authorize or block the notification process.
     * If the boolean or the callable return false, no notification will be sent.
     */
-    'processAllowedToRun' => true,
+    'allowedToRun' => true,
     // could also be :
-    // 'processAllowedToRun' => function () {
+    // 'allowedToRun' => function () {
     //     return true;
     // },