+++ src/Exceptions/InvalidAllowedToRun.php
@@ -4,7 +4,7 @@ namespace Okipa\LaravelFailedJobsNotifier\Exceptions;
 
 use Exception;
 
-class InvalidProcessAllowedToRun extends Exception
+class InvalidAllowedToRun extends Exception
 {
     //
 }